August-2019-Commit - Changes -- Save_images.ijm Deleted need for "Images" folder from documentation. The change in the code dates already a while back.
--Align_script.ijm
Now does a maximum intensity projection of the whole stack if there are more than 1 z-slices. This leads to easier time for alignment.

-- Analyzing_enrichment.ijm
Added a call to "Set measurements" to automatically choose Label, Area and Mean Gray Value. These values are needed for a correct analysis later on. This fix solves the problems of some users that did not set the measurements beforehand.